+++ app/form_objects/notification_form.rb
@@ -1,0 +1,34 @@
+# Represents form object for Notification model
+class NotificationForm
+  include ActiveModel::Model
+  include Virtus.model
+
+  attribute :name, Hash
+  attribute :event, String
+
+  validates :name, :event, presence: true
+  validate :name_as_hash
+
+  attr_reader :notification
+
+  def persist?
+    return false unless valid?
+    return false if exists?
+    @notification = Notification.new
+    @notification.attributes = attributes
+    @notification.save
+    true
+  end
+
+  private
+
+  def exists?
+    Notification.find_by(name: name).present?
+  end
+
+  def name_as_hash
+    return errors[:name] << 'Name is not hash' unless name.is_a?(Hash)
+    errors[:name] << 'Name EN is empty' unless name['en'].present?
+    errors[:name] << 'Name RU is empty' unless name['ru'].present?
+  end
+end