+++ app/models/notification.rb
@@ -1,0 +1,3 @@
+# Represents notifications about events
+class Notification < ApplicationRecord
+end