+++ spec/factories/notifications.rb
@@ -1,0 +1,6 @@
+FactoryBot.define do
+  factory :notification do
+    sequence(:name) { |i| "Name#{i}" }
+    event { 'guild_event_creation' }
+  end
+end